Move `reset`
